A little spit (and polish)
